// src/app/models/venue-editor-seat.ts
export class VenueEditorSeat {
    BlockId?: number;
    BlockName: string;
    CustomerId?: string;
    GateId?: number;
    Id: number;
    IsStanding?: boolean;
    Name?: string;
    ProductId?: number;
    VariantId?: number;
    RowId?: number;
    RowName?: string;
    SeatColor?: string;
    SeatPriority?: number;
    Status?: number;
    Status_Desc?: string;
    TicketType?: number;
    TicketType_Desc?: string;
    Type?: number;
    Type_Desc?: string;
}